hpn #014 move unit tests to folders
