+++ sources/tests/Eshva.DockerCompose.Tests.Unit/Commands/Down/GivenDownProjectCommandWhenExecuteCalled.cs
@@ -3,7 +3,6 @@
 using System;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using Eshva.DockerCompose.Commands.UpProject;
 using Eshva.DockerCompose.Infrastructure;
 using Moq;
 using Xunit;
@@ -11,20 +10,19 @@ using Xunit;
 #endregion
 
 
-namespace Eshva.DockerCompose.Tests.Unit.Commands
+namespace Eshva.DockerCompose.Tests.Unit.Commands.Down
 {
-    public sealed class GivenUpProjectCommandWhenExecuteCalled
+    public sealed class GivenDownProjectCommandWhenExecuteCalled
     {
         [Fact]
-        public async Task ShouldContainValidArgumentsIfDetached()
+        public async Task ShouldContainValidArguments()
         {
             var starterMock = new Mock<IProcessStarter>();
             Expression<Func<string, bool>> argumentsValidator =
-                arguments => arguments.Equals("-f \"project.yaml\" up --detach", StringComparison.OrdinalIgnoreCase);
+                arguments => arguments.Equals("-f \"project.yaml\" down", StringComparison.OrdinalIgnoreCase);
             starterMock.Setup(starter => starter.Start(It.Is(argumentsValidator), It.IsAny<TimeSpan>()))
                        .Returns(Task.FromResult(0));
-            var command = UpProjectCommand.WithFilesAndStarter(starterMock.Object, "project.yaml")
-                                          .Build();
+            var command = new DockerCompose.Commands.DownProject.DownProjectCommand(starterMock.Object, "project.yaml");
             await command.Execute();
             starterMock.Verify(
                 starter => starter.Start(It.Is(argumentsValidator), It.IsAny<TimeSpan>()),
